# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AddSubRequest:
    properties:
      price:
        type: integer
      service_name:
        type: string
      user_id:
        type: string
    type: object
  models.EditSubRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Subs:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  subscriptions.AddSubscriptionResponse:
    properties:
      id:
        type: string
    type: object
  subscriptions.EditSubscriptionResponse:
    properties:
      status:
        type: string
    type: object
  subscriptions.GetPriceWithPeriodResponse:
    properties:
      price:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation example
  title: Online Subscriptions Swagger Api
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Get subscriptions from service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subs'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Add subscription to service
      parameters:
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddSubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscriptions.AddSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete subscription from service
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid ID parameter
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Remove subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Get subscription from service
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get subscription
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Edit subscription in service
      parameters:
      - description: Subscription ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EditSubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.EditSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Edit subscription
      tags:
      - subscriptions
  /subscriptions/price:
    get:
      consumes:
      - application/json
      description: Get price with period from service
      parameters:
      - description: Start date
        example: "2025-01-01"
        format: date
        in: query
        name: from
        required: true
        type: string
      - description: End date
        example: "2025-01-31"
        format: date
        in: query
        name: to
        required: true
        type: string
      - description: User ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: query
        name: user_id
        required: true
        type: string
      - description: Service name
        example: '"premium"'
        in: query
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.GetPriceWithPeriodResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get price
      tags:
      - subscriptions
swagger: "2.0"